block firestore CRUD actions if not signed in

+++ lib/firework/firestore_crub_helper.dart
@@ -1,4 +1,5 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 
 abstract class DocumentBase {
   String get collectionPath;
@@ -8,16 +9,22 @@ abstract class DocumentBase {
 
 mixin FirestoreCRUD on DocumentBase {
   void createData() {
+    if (FirebaseAuth.instance.currentUser == null) return;
+
     final ref = FirebaseFirestore.instance.collection(collectionPath).doc();
     ref.set(toJson());
   }
 
   void deleteData(String id) {
+    if (FirebaseAuth.instance.currentUser == null) return;
+
     final ref = FirebaseFirestore.instance.collection(collectionPath).doc(id);
     ref.delete();
   }
 
   void updateData(String id) {
+    if (FirebaseAuth.instance.currentUser == null) return;
+
     final ref = FirebaseFirestore.instance.collection(collectionPath).doc(id);
     ref.update(toJson());
   }
